Update `obtainedMark` type from `Long` to `Float` across models
Changed the data type of `obtainedMark` in DTOs, entities, and repositories to allow decimal values. Updated related methods, mappings, and records to ensure compatibility. These changes improve precision for scoring calculations.

+++ src/main/java/com/edusyspro/api/dto/custom/ScoreBasic.java
@@ -10,7 +10,7 @@ public record ScoreBasic(
         Long scoreId,
         UUID studentId,
         Long assignmentId,
-        Long obtainedMark
+        Float obtainedMark
 ) {
     public ScoreDTO toDTO() {
         return ScoreDTO.builder()
