+++ src/main/java/com/edusyspro/api/dto/custom/ScoreBasicValue.java
@@ -16,7 +16,7 @@ public record ScoreBasicValue(
         String studentImage,
         String studentReference,
         String classeName,
-        Long obtainedMark,
+        Float obtainedMark,
         Boolean isPresent
 ) {
